requirements.txt
Django>=3.1, <3.2
asgiref~=3.2.10
beautifulsoup4~=4.9.3
celery~=4.4.2
django-ckeditor==5.9.0
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.8
django-filter==2.3.0
django-json-ld
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
django-rest-swagger==2.2.0
django-taggit==1.3.0
django-taggit-autosuggest==0.3.8
django-redis
channels==2.4.0
channels-redis==3.1.0
hazm==0.7.0
Khayyam==3.0.17
pid~=3.0.4
Pillow==7.2.0
pika~=1.1.0
psycopg2-binary
python-decouple==3.3
python-memcached==1.59
requests==2.24.0
sentry-sdk==0.19.1
sorl-thumbnail==12.6.3
